+++ magda-int-test/src/test/scala/au/csiro/data61/magda/test/util/MagdaMatchers.scala
@@ -4,6 +4,8 @@ import au.csiro.data61.magda.model.misc.DataSet
 import org.tartarus.snowball.ext.PorterStemmer
 import org.apache.lucene.analysis.standard.StandardTokenizer
 import org.apache.lucene.analysis.tokenattributes.CharTermAttribute
+import org.apache.lucene.analysis.en._
+import au.csiro.data61.magda.test.util.LightEnglishStemmer
 
 object MagdaMatchers extends org.scalatest.Matchers {
   def dataSetEqual(ds1: DataSet, ds2: DataSet) = ds1.copy(
@@ -24,6 +26,11 @@ object MagdaMatchers extends org.scalatest.Matchers {
     stemmer.setCurrent(string)
 
     if (stemmer.stem) stemmer.getCurrent else string
+  }
+
+  def lightEnglishStem(string: String): String = {
+    val stemmer = new LightEnglishStemmer
+    return stemmer.stem(string)
   }
 
   def toEnglishToken(string: String) = porterStem(string.toLowerCase)