+++ magda-search-api/src/main/scala/au/csiro/data61/magda/search/elasticsearch/ElasticSearchQueryer.scala
@@ -70,7 +70,7 @@ class ElasticSearchQueryer(indices: Indices = DefaultIndices)(
       throw t
   }
 
-  val DATASETS_LANGUAGE_FIELDS = Seq(("title", 20f), ("description", 5f), "publisher.name", ("keywords", 5f), "themes")
+  val DATASETS_LANGUAGE_FIELDS = Seq(("title", 50f), ("description", 2f), "publisher.name", ("keywords", 10f), "themes")
   val NON_LANGUAGE_FIELDS = Seq("_id", "catalog", "accrualPeriodicity", "contactPoint.name", "publisher.acronym")
 
   override def search(inputQuery: Query, start: Long, limit: Int, requestedFacetSize: Int) = {
@@ -321,13 +321,8 @@ class ElasticSearchQueryer(indices: Indices = DefaultIndices)(
     val clauses: Seq[Traversable[QueryDefinition]] = Seq(
       query.freeText flatMap { inputText =>
         val text = if (inputText.trim.length == 0) "*" else inputText
-        val queryString = new SimpleStringQueryDefinition(text).defaultOperator(operator).quoteFieldSuffix("keyword")
+        val queryString = new SimpleStringQueryDefinition(text).defaultOperator(operator).quoteFieldSuffix(".quote")
 
-        // For some reason to make english analysis work properly you need to specifically hit the english fields.
-        def foldFieldsEnglish(query: SimpleStringQueryDefinition, fields: Seq[Any]) = fields.foldRight(query) {
-          case ((fieldName: String, boost: Float), queryDef) => queryDef.field(fieldName + ".english", boost)
-          case (field: String, queryDef)                     => queryDef.field(field + ".english", 0)
-        }
         def foldFields(query: SimpleStringQueryDefinition, fields: Seq[Any]) = fields.foldRight(query) {
           case ((fieldName: String, boost: Float), queryDef) => queryDef.field(fieldName, boost)
           case (field: String, queryDef)                     => queryDef.field(field, 0)
@@ -339,11 +334,11 @@ class ElasticSearchQueryer(indices: Indices = DefaultIndices)(
           .query(
             // If this was AND then a single distribution would have to match the entire query, this way you can
             // have multiple dists partially match
-            queryString.field("distributions.title.english").field("distributions.description.english").field("distributions.format.keyword_lowercase")
+            queryString.field("distributions.title").field("distributions.description").field("distributions.format.keyword_lowercase")
               .defaultOperator("or"))
           .scoreMode(ScoreMode.Max)
 
-        val queries = Seq(foldFieldsEnglish(foldFields(queryString, NON_LANGUAGE_FIELDS), DATASETS_LANGUAGE_FIELDS), distributionsEnglishQueries)
+        val queries = Seq(foldFields(queryString, NON_LANGUAGE_FIELDS++DATASETS_LANGUAGE_FIELDS), distributionsEnglishQueries)
 
         Some(dismax(queries).tieBreaker(0.2))
       },
@@ -365,7 +360,7 @@ class ElasticSearchQueryer(indices: Indices = DefaultIndices)(
           .defaultOperator("or")
           .analyzeWildcard(true)
           .field("_all")
-          .field("value.english"))
+          .field("value"))
         .start(start.toInt)
         .limit(limit))
         .flatMap { response =>